requirements.txt
azure-iot-device
CounterFit
counterfit-connection
counterfit-shims-grove
pytest
mockito
pytest-cov
radon
pytest-dotenv
flake8